// formbuilder-frontend1/src/types/schema.ts
export type FieldType = 'TEXT' | 'NUMERIC' | 'DATE' | 'BOOLEAN' | 'TEXTAREA'
  | 'DROPDOWN' | 'RADIO' | 'CHECKBOX_GROUP'
  | 'TIME' | 'RATING' | 'SCALE' | 'FILE' | 'GRID_RADIO' | 'GRID_CHECK' | 'LOOKUP'
  | 'CALCULATED' | 'SECTION_HEADER' | 'INFO_LABEL' | 'PAGE_BREAK' | 'DATE_TIME' | 'HIDDEN';